requirement.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.5
cycler==0.11.0
dataclasses==0.6
einops==0.4.1
filelock==3.4.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py~=3.1.0
huggingface-hub==0.4.0
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
keras-preprocessing==1.1.2
kiwisolver==1.3.1
markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.2
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywavelets==1.1.1
pyyaml==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.15.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.6.12
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.64.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.18
werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
